+++ library/src/main/java/com/airbnb/android/airmapview/AirMapMarker.java
@@ -1,123 +1,178 @@
 package com.airbnb.android.airmapview;
 
-import com.google.android.gms.maps.GoogleMap;
+import android.graphics.Bitmap;
+
+import com.google.android.gms.maps.model.BitmapDescriptor;
 import com.google.android.gms.maps.model.BitmapDescriptorFactory;
 import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.Marker;
 import com.google.android.gms.maps.model.MarkerOptions;
 
-import android.graphics.Bitmap;
-import android.text.TextUtils;
-
 /**
- * Helper class for keeping record of data needed to display map markers, as well as an object T
- * associated with the marker
+ * Wrapper around {@link MarkerOptions}. Keeps record of data needed to display map markers, as
+ * well as an object T associated with the marker.
  */
 public class AirMapMarker<T> {
 
-  private T object;
-  private long id;
-  private LatLng latLng;
-  private String title;
-  private int iconId;
-  private Bitmap bitmap;
-  private Marker googleMarker;
+  private final T object;
+  private final long id;
+  private final MarkerOptions markerOptions;
+  private Marker marker;
 
-  public AirMapMarker(LatLng latLng, long id) {
-    this(null, latLng, id);
+  private AirMapMarker(T object, long id, MarkerOptions markerOptions) {
+    this.object = object;
+    this.id = id;
+    this.markerOptions = markerOptions;
   }
 
-  public AirMapMarker(T object, LatLng latLng, long id) {
-    this.object = object;
-    this.latLng = latLng;
-    this.id = id;
+  public T object() {
+    return object;
   }
 
   public long getId() {
     return id;
   }
 
-  public AirMapMarker<T> setId(long id) {
-    this.id = id;
-    return this;
-  }
-
   public LatLng getLatLng() {
-    return latLng;
+    return markerOptions.getPosition();
   }
 
-  public AirMapMarker<T> setLatLng(LatLng latLng) {
-    this.latLng = latLng;
-    return this;
+  void setLatLng(LatLng latLng) {
+    markerOptions.position(latLng);
   }
 
   public String getTitle() {
-    return title;
+    return markerOptions.getTitle();
   }
 
-  public AirMapMarker<T> setTitle(String title) {
-    this.title = title;
-    return this;
+  public String getSnippet() {
+    return markerOptions.getSnippet();
   }
 
-  public int getIconId() {
-    return iconId;
+  public MarkerOptions getMarkerOptions() {
+    return markerOptions;
   }
 
-  public AirMapMarker<T> setIconId(int iconId) {
-    this.iconId = iconId;
-    return this;
+  /** Sets a marker associated to this object */
+  void setGoogleMarker(Marker marker) {
+    this.marker = marker;
   }
 
-  public AirMapMarker<T> setIcon(Bitmap bitmap) {
-    this.bitmap = bitmap;
-    return this;
+  public Builder<T> toBuilder() {
+    return new Builder<T>()
+        .id(id)
+        .object(object)
+        .position(markerOptions.getPosition())
+        .alpha(markerOptions.getAlpha())
+        .anchor(markerOptions.getAnchorU(), markerOptions.getAnchorV())
+        .bitmapDescriptor(markerOptions.getIcon())
+        .infoWindowAnchor(markerOptions.getInfoWindowAnchorU(), markerOptions.getInfoWindowAnchorV())
+        .snippet(markerOptions.getSnippet())
+        .title(markerOptions.getTitle())
+        .draggable(markerOptions.isDraggable())
+        .visible(markerOptions.isVisible())
+        .alpha(markerOptions.getAlpha())
+        .rotation(markerOptions.getRotation())
+        .flat(markerOptions.isFlat());
+  }
+  
+  public Marker getMarker() {
+    return marker;
   }
 
-  public T getObject() {
-    return object;
-  }
+  public static class Builder<T> {
+    private T object;
+    private long id;
+    private final MarkerOptions markerOptions = new MarkerOptions();
 
-  public AirMapMarker<T> setObject(T object) {
-    this.object = object;
-    return this;
-  }
-
-  /**
-   * Add this marker to the given {@link com.google.android.gms.maps.GoogleMap} instance
-   *
-   * @param googleMap the {@link com.google.android.gms.maps.GoogleMap} instance to which the marker
-   *                  will be added
-   */
-  public void addToGoogleMap(GoogleMap googleMap) {
-    MarkerOptions options = new MarkerOptions();
-
-    options.position(latLng);
-
-    if (!TextUtils.isEmpty(title)) {
-      options.title(title);
+    public Builder() {
     }
 
-    if (bitmap != null) {
-      options.icon(BitmapDescriptorFactory.fromBitmap(bitmap));
-    } else if (iconId > 0) {
-      options.icon(BitmapDescriptorFactory.fromResource(iconId));
+    public Builder<T> object(T object) {
+      this.object = object;
+      return this;
     }
 
-    // add the marker and keep a reference so it can be removed
-    googleMarker = googleMap.addMarker(options);
-  }
-
-  /**
-   * Remove this polyline from a GoogleMap (if it was added).
-   *
-   * @return true if the {@link com.google.android.gms.maps.model.Polyline} was removed
-   */
-  public boolean removeFromGoogleMap() {
-    if (googleMarker != null) {
-      googleMarker.remove();
-      return true;
+    public Builder<T> id(long id) {
+      this.id = id;
+      return this;
     }
-    return false;
+
+    public Builder<T> position(LatLng position) {
+      markerOptions.position(position);
+      return this;
+    }
+
+    public Builder<T> anchor(float u, float v) {
+      markerOptions.anchor(u, v);
+      return this;
+    }
+
+    public Builder<T> infoWindowAnchor(float u, float v) {
+      markerOptions.infoWindowAnchor(u, v);
+      return this;
+    }
+
+    public Builder<T> title(String title) {
+      markerOptions.title(title);
+      return this;
+    }
+
+    public Builder<T> snippet(String snippet) {
+      markerOptions.snippet(snippet);
+      return this;
+    }
+
+    public Builder<T> iconId(int iconId) {
+      try {
+        markerOptions.icon(BitmapDescriptorFactory.fromResource(iconId));
+      } catch (NullPointerException ignored) {
+        // google play services is not available
+      }
+      return this;
+    }
+
+    public Builder<T> bitmap(Bitmap bitmap) {
+      try {
+        bitmapDescriptor(BitmapDescriptorFactory.fromBitmap(bitmap));
+      } catch (NullPointerException ignored) {
+        // google play services is not available
+      }
+      return this;
+    }
+
+    public Builder<T> bitmapDescriptor(BitmapDescriptor bitmap) {
+      markerOptions.icon(bitmap);
+      return this;
+    }
+
+    public Builder<T> draggable(boolean draggable) {
+      markerOptions.draggable(draggable);
+      return this;
+    }
+
+    public Builder<T> visible(boolean visible) {
+      markerOptions.visible(visible);
+      return this;
+    }
+
+    public Builder<T> flat(boolean flat) {
+      markerOptions.flat(flat);
+      return this;
+    }
+
+    public Builder<T> rotation(float rotation) {
+      markerOptions.rotation(rotation);
+      return this;
+    }
+
+    public Builder<T> alpha(float alpha) {
+      markerOptions.alpha(alpha);
+      return this;
+    }
+
+    public AirMapMarker<T> build() {
+      return new AirMapMarker<>(object, id, markerOptions);
+    }
   }
 }