+++ library/src/main/java/com/airbnb/android/airmapview/GoogleChinaWebViewMapFragment.java
@@ -1,9 +1,13 @@
 package com.airbnb.android.airmapview;
 
-public class GoogleChinaWebViewMapFragment extends WebViewMapFragment {
-
+public class GoogleChinaWebViewMapFragment extends GoogleWebViewMapFragment {
   public static GoogleChinaWebViewMapFragment newInstance(AirMapType mapType) {
     return (GoogleChinaWebViewMapFragment) new GoogleChinaWebViewMapFragment()
         .setArguments(mapType);
+  }
+
+  @Override
+  protected boolean isChinaMode() {
+    return true;
   }
 }