+++ library/src/main/java/com/airbnb/android/airmapview/WebViewMapFragment.java
@@ -1,10 +1,7 @@
 package com.airbnb.android.airmapview;
 
-import com.google.android.gms.maps.GoogleMap;
-import com.google.android.gms.maps.model.LatLng;
-import com.google.android.gms.maps.model.LatLngBounds;
-
 import android.annotation.SuppressLint;
+import android.graphics.Bitmap;
 import android.graphics.Point;
 import android.os.Bundle;
 import android.os.Handler;
@@ -12,14 +9,14 @@ import android.os.Looper;
 import android.support.annotation.NonNull;
 import android.support.v4.app.Fragment;
 import android.util.Log;
-import android.view.Gravity;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.webkit.GeolocationPermissions;
 import android.webkit.JavascriptInterface;
+import android.webkit.WebChromeClient;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
-import android.widget.FrameLayout;
 
 import com.airbnb.android.airmapview.listeners.InfoWindowCreator;
 import com.airbnb.android.airmapview.listeners.OnCameraChangeListener;
@@ -29,23 +26,30 @@ import com.airbnb.android.airmapview.listeners.OnMapBoundsCallback;
 import com.airbnb.android.airmapview.listeners.OnMapClickListener;
 import com.airbnb.android.airmapview.listeners.OnMapLoadedListener;
 import com.airbnb.android.airmapview.listeners.OnMapMarkerClickListener;
+import com.airbnb.android.airmapview.listeners.OnMapMarkerDragListener;
+import com.airbnb.android.airmapview.listeners.OnSnapshotReadyListener;
+import com.google.android.gms.maps.GoogleMap;
+import com.google.android.gms.maps.model.LatLng;
+import com.google.android.gms.maps.model.LatLngBounds;
 
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import java.util.List;
+import java.util.HashMap;
+import java.util.Locale;
+import java.util.Map;
 
 public abstract class WebViewMapFragment extends Fragment implements AirMapInterface {
-
   private static final String TAG = WebViewMapFragment.class.getSimpleName();
 
-  private WebView webView;
+  protected WebView webView;
   private ViewGroup mLayout;
   private OnMapClickListener onMapClickListener;
   private OnCameraChangeListener onCameraChangeListener;
   private OnMapLoadedListener onMapLoadedListener;
   private OnMapMarkerClickListener onMapMarkerClickListener;
+  private OnMapMarkerDragListener onMapMarkerDragListener;
   private OnInfoWindowClickListener onInfoWindowClickListener;
   private InfoWindowCreator infoWindowCreator;
   private OnMapBoundsCallback onMapBoundsCallback;
@@ -55,15 +59,27 @@ public abstract class WebViewMapFragment extends Fragment implements AirMapInter
   private boolean loaded;
   private boolean ignoreNextMapMove;
   private View infoWindowView;
+  private final Map<Long, AirMapMarker<?>> markers = new HashMap<>();
+
+  private boolean trackUserLocation = false;
 
   public WebViewMapFragment setArguments(AirMapType mapType) {
     setArguments(mapType.toBundle());
     return this;
   }
 
-  @SuppressLint({"SetJavaScriptEnabled", "AddJavascriptInterface"})
-  @Override
-  public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+  public class GeoWebChromeClient extends WebChromeClient {
+    @Override public void onGeolocationPermissionsShowPrompt(
+        String origin, GeolocationPermissions.Callback callback) {
+      // Always grant permission since the app itself requires location
+      // permission and the user has therefore already granted it
+      callback.invoke(origin, true, false);
+    }
+  }
+
+  @SuppressLint({ "SetJavaScriptEnabled", "AddJavascriptInterface" })
+  @Override public View onCreateView(
+      LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
     View view = inflater.inflate(R.layout.fragment_webview, container, false);
 
     webView = (WebView) view.findViewById(R.id.webview);
@@ -73,11 +89,15 @@ public abstract class WebViewMapFragment extends Fragment implements AirMapInter
     webViewSettings.setSupportZoom(true);
     webViewSettings.setBuiltInZoomControls(false);
     webViewSettings.setJavaScriptEnabled(true);
+    webViewSettings.setGeolocationEnabled(true);
+    webViewSettings.setAllowFileAccess(false);
+    webViewSettings.setAllowContentAccess(false);
+    webView.setWebChromeClient(new GeoWebChromeClient());
 
     AirMapType mapType = AirMapType.fromBundle(getArguments());
-
     webView.loadDataWithBaseURL(mapType.getDomain(), mapType.getMapData(getResources()),
-                                 "text/html", "base64", null);
+        "text/html", "base64", null);
+
     webView.addJavascriptInterface(new MapsJavaScriptInterface(), "AirMapView");
 
     return view;
@@ -92,8 +112,8 @@ public abstract class WebViewMapFragment extends Fragment implements AirMapInter
   }
 
   public void setCenter(LatLng latLng) {
-    webView.loadUrl(String.format("javascript:centerMap(%1$f, %2$f);", latLng.latitude,
-                                  latLng.longitude));
+    webView.loadUrl(String.format(Locale.US, "javascript:centerMap(%1$f, %2$f);", latLng.latitude,
+        latLng.longitude));
   }
 
   public void animateCenter(LatLng latLng) {
@@ -101,7 +121,7 @@ public abstract class WebViewMapFragment extends Fragment implements AirMapInter
   }
 
   public void setZoom(int zoom) {
-    webView.loadUrl(String.format("javascript:setZoom(%1$d);", zoom));
+    webView.loadUrl(String.format(Locale.US, "javascript:setZoom(%1$d);", zoom));
   }
 
   public void drawCircle(LatLng latLng, int radius) {
@@ -117,21 +137,22 @@ public abstract class WebViewMapFragment extends Fragment implements AirMapInter
   }
 
   @Override public void drawCircle(LatLng latLng, int radius, int borderColor, int borderWidth,
-                         int fillColor) {
-    webView.loadUrl(String.format("javascript:addCircle(%1$f, %2$f, %3$d, %4$d, %5$d, %6$d);",
-                                  latLng.latitude, latLng.longitude, radius, borderColor,
-                                  borderWidth, fillColor));
+      int fillColor) {
+    webView.loadUrl(
+        String.format(Locale.US, "javascript:addCircle(%1$f, %2$f, %3$d, %4$d, %5$d, %6$d);",
+            latLng.latitude, latLng.longitude, radius, borderColor,
+            borderWidth, fillColor));
   }
 
   public void highlightMarker(long markerId) {
     if (markerId != -1) {
-      webView.loadUrl(String.format("javascript:highlightMarker(%1$d);", markerId));
+      webView.loadUrl(String.format(Locale.US, "javascript:highlightMarker(%1$d);", markerId));
     }
   }
 
   public void unhighlightMarker(long markerId) {
     if (markerId != -1) {
-      webView.loadUrl(String.format("javascript:unhighlightMarker(%1$d);", markerId));
+      webView.loadUrl(String.format(Locale.US, "javascript:unhighlightMarker(%1$d);", markerId));
     }
   }
 
@@ -139,17 +160,30 @@ public abstract class WebViewMapFragment extends Fragment implements AirMapInter
     return webView != null && loaded;
   }
 
-  @Override public void addMarker(AirMapMarker marker) {
+  @Override public void addMarker(AirMapMarker<?> marker) {
     LatLng latLng = marker.getLatLng();
-    webView.loadUrl(String.format("javascript:addMarkerWithId(%1$f, %2$f, %3$d);",
-                                   latLng.latitude, latLng.longitude, marker.getId()));
+    markers.put(marker.getId(), marker);
+    webView.loadUrl(
+        String.format(Locale.US,
+            "javascript:addMarkerWithId(%1$f, %2$f, %3$d, '%4$s', '%5$s', %6$b);",
+            latLng.latitude, latLng.longitude, marker.getId(), marker.getTitle(),
+            marker.getSnippet(), marker.getMarkerOptions().isDraggable()));
   }
 
-  @Override public void removeMarker(AirMapMarker marker) {
-    webView.loadUrl(String.format("javascript:removeMarker(%1$d);", marker.getId()));
+  @Override public void moveMarker(AirMapMarker<?> marker, LatLng to) {
+    marker.setLatLng(to);
+    webView.loadUrl(
+        String.format(Locale.US, "javascript:moveMarker(%1$f, %2$f, '%3$s');",
+            to.latitude, to.longitude, marker.getId()));
+  }
+
+  @Override public void removeMarker(AirMapMarker<?> marker) {
+    markers.remove(marker.getId());
+    webView.loadUrl(String.format(Locale.US, "javascript:removeMarker(%1$d);", marker.getId()));
   }
 
   public void clearMarkers() {
+    markers.clear();
     webView.loadUrl("javascript:clearMarkers();");
   }
 
@@ -177,18 +211,45 @@ public abstract class WebViewMapFragment extends Fragment implements AirMapInter
     onMapMarkerClickListener = listener;
   }
 
+  @Override public void setOnMarkerDragListener(OnMapMarkerDragListener listener) {
+    onMapMarkerDragListener = listener;
+  }
+
   @Override public void setPadding(int left, int top, int right, int bottom) {
     // no-op
   }
 
-  @Override public void setMyLocationEnabled(boolean b) {
+  @Override public void setMyLocationEnabled(boolean trackUserLocationEnabled) {
+    trackUserLocation = trackUserLocationEnabled;
+    if (trackUserLocationEnabled) {
+      RuntimePermissionUtils.checkLocationPermissions(getActivity(), this);
+    } else {
+      webView.loadUrl("javascript:stopTrackingUserLocation();");
+    }
+  }
+
+  @Override public void onLocationPermissionsGranted() {
+    webView.loadUrl("javascript:startTrackingUserLocation();");
+  }
+
+  @Override public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions,
+      @NonNull int[] grantResults) {
+    super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+    RuntimePermissionUtils.onRequestPermissionsResult(this, requestCode, grantResults);
+  }
+
+  @Override public boolean isMyLocationEnabled() {
+    return trackUserLocation;
+  }
+
+  @Override public void setMapToolbarEnabled(boolean enabled) {
     // no-op
   }
 
-  @Override public void addPolyline(AirMapPolyline polyline) {
+  @Override public <T> void addPolyline(AirMapPolyline<T> polyline) {
     try {
       JSONArray array = new JSONArray();
-      for (LatLng point : (List<LatLng>) polyline.getPoints()) {
+      for (LatLng point : polyline.getPoints()) {
         JSONObject json = new JSONObject();
         json.put("lat", point.latitude);
         json.put("lng", point.longitude);
@@ -203,8 +264,33 @@ public abstract class WebViewMapFragment extends Fragment implements AirMapInter
     }
   }
 
-  @Override public void removePolyline(AirMapPolyline polyline) {
-    webView.loadUrl(String.format("javascript:removePolyline(%1$d);", polyline.getId()));
+  @Override public <T> void removePolyline(AirMapPolyline<T> polyline) {
+    webView.loadUrl(String.format(Locale.US, "javascript:removePolyline(%1$d);", polyline.getId()));
+  }
+
+  @Override public <T> void addPolygon(AirMapPolygon<T> polygon) {
+    try {
+      JSONArray array = new JSONArray();
+      for (LatLng point : polygon.getPolygonOptions().getPoints()) {
+        JSONObject json = new JSONObject();
+        json.put("lat", point.latitude);
+        json.put("lng", point.longitude);
+        array.put(json);
+      }
+
+      webView.loadUrl(String.format(Locale.US,
+          "javascript:addPolygon(" + array.toString() + ", %1$d, %2$d, %3$d, %4$d);",
+          polygon.getId(),
+          (int) polygon.getPolygonOptions().getStrokeWidth(),
+          polygon.getPolygonOptions().getStrokeColor(),
+          polygon.getPolygonOptions().getFillColor()));
+    } catch (JSONException e) {
+      Log.e(TAG, "error constructing polyline JSON", e);
+    }
+  }
+
+  @Override public <T> void removePolygon(AirMapPolygon<T> polygon) {
+    webView.loadUrl(String.format(Locale.US, "javascript:removePolygon(%1$d);", polygon.getId()));
   }
 
   @Override public void setOnMapClickListener(final OnMapClickListener listener) {
@@ -216,7 +302,7 @@ public abstract class WebViewMapFragment extends Fragment implements AirMapInter
   }
 
   @Override public void setInfoWindowCreator(GoogleMap.InfoWindowAdapter adapter,
-                                   InfoWindowCreator creator) {
+      InfoWindowCreator creator) {
     infoWindowCreator = creator;
   }
 
@@ -227,25 +313,31 @@ public abstract class WebViewMapFragment extends Fragment implements AirMapInter
 
   @Override public void getScreenLocation(LatLng latLng, OnLatLngScreenLocationCallback callback) {
     onLatLngScreenLocationCallback = callback;
-    webView.loadUrl(String.format("javascript:getScreenLocation(%1$f, %2$f);",
-                                  latLng.latitude, latLng.longitude));
+    webView.loadUrl(String.format(Locale.US, "javascript:getScreenLocation(%1$f, %2$f);",
+        latLng.latitude, latLng.longitude));
   }
 
   @Override public void setCenter(LatLngBounds bounds, int boundsPadding) {
-    webView.loadUrl(String.format("javascript:setBounds(%1$f, %2$f, %3$f, %4$f);",
-                                   bounds.northeast.latitude, bounds.northeast.longitude,
-                                   bounds.southwest.latitude,
-                                   bounds.southwest.longitude));
+    webView.loadUrl(String.format(Locale.US, "javascript:setBounds(%1$f, %2$f, %3$f, %4$f);",
+        bounds.northeast.latitude, bounds.northeast.longitude,
+        bounds.southwest.latitude,
+        bounds.southwest.longitude));
+  }
+
+  protected boolean isChinaMode() {
+    return false;
   }
 
   private class MapsJavaScriptInterface {
-
     private final Handler handler = new Handler(Looper.getMainLooper());
+
+    @JavascriptInterface public boolean isChinaMode() {
+      return WebViewMapFragment.this.isChinaMode();
+    }
 
     @JavascriptInterface public void onMapLoaded() {
       handler.post(new Runnable() {
-        @Override
-        public void run() {
+        @Override public void run() {
           if (!loaded) {
             loaded = true;
             if (onMapLoadedListener != null) {
@@ -258,8 +350,7 @@ public abstract class WebViewMapFragment extends Fragment implements AirMapInter
 
     @JavascriptInterface public void mapClick(final double lat, final double lng) {
       handler.post(new Runnable() {
-        @Override
-        public void run() {
+        @Override public void run() {
           if (onMapClickListener != null) {
             onMapClickListener.onMapClick(new LatLng(lat, lng));
           }
@@ -270,12 +361,12 @@ public abstract class WebViewMapFragment extends Fragment implements AirMapInter
       });
     }
 
-    @JavascriptInterface public void getBoundsCallback(double neLat, double neLng, double swLat, double swLng) {
+    @JavascriptInterface public void getBoundsCallback(
+        double neLat, double neLng, double swLat, double swLng) {
       final LatLngBounds bounds = new LatLngBounds(new LatLng(swLat, swLng),
-                                                   new LatLng(neLat, neLng));
+          new LatLng(neLat, neLng));
       handler.post(new Runnable() {
-        @Override
-        public void run() {
+        @Override public void run() {
           onMapBoundsCallback.onMapBoundsReady(bounds);
         }
       });
@@ -284,8 +375,7 @@ public abstract class WebViewMapFragment extends Fragment implements AirMapInter
     @JavascriptInterface public void getLatLngScreenLocationCallback(int x, int y) {
       final Point point = new Point(x, y);
       handler.post(new Runnable() {
-        @Override
-        public void run() {
+        @Override public void run() {
           onLatLngScreenLocationCallback.onLatLngScreenLocationReady(point);
         }
       });
@@ -296,8 +386,7 @@ public abstract class WebViewMapFragment extends Fragment implements AirMapInter
       WebViewMapFragment.this.zoom = zoom;
 
       handler.post(new Runnable() {
-        @Override
-        public void run() {
+        @Override public void run() {
           if (onCameraChangeListener != null) {
             onCameraChangeListener.onCameraChanged(center, WebViewMapFragment.this.zoom);
           }
@@ -314,12 +403,12 @@ public abstract class WebViewMapFragment extends Fragment implements AirMapInter
       });
     }
 
-    @JavascriptInterface public void markerClick(final long markerId) {
+    @JavascriptInterface public void markerClick(long markerId) {
+      final AirMapMarker<?> airMapMarker = markers.get(markerId);
       handler.post(new Runnable() {
-        @Override
-        public void run() {
+        @Override public void run() {
           if (onMapMarkerClickListener != null) {
-            onMapMarkerClickListener.onMapMarkerClick(markerId);
+            onMapMarkerClickListener.onMapMarkerClick(airMapMarker);
           }
 
           if (infoWindowView != null) {
@@ -328,28 +417,100 @@ public abstract class WebViewMapFragment extends Fragment implements AirMapInter
 
           // TODO convert to custom dialog fragment
           if (infoWindowCreator != null) {
-            infoWindowView = infoWindowCreator.createInfoWindow(markerId);
-            int height = (int) getResources().getDimension(R.dimen.map_marker_height);
-            FrameLayout.LayoutParams layoutParams = new FrameLayout.LayoutParams(
-                FrameLayout.LayoutParams.WRAP_CONTENT, height, Gravity.CENTER);
-            layoutParams.bottomMargin = height;
-            infoWindowView.setLayoutParams(layoutParams);
-            mLayout.addView(infoWindowView);
-
-            infoWindowView.setOnClickListener(new View.OnClickListener() {
-              @Override
-              public void onClick(@NonNull View v) {
-                mLayout.removeView(infoWindowView);
-                if (onInfoWindowClickListener != null) {
-                  onInfoWindowClickListener.onInfoWindowClick(markerId);
+            infoWindowView = infoWindowCreator.createInfoWindow(airMapMarker);
+            if (infoWindowView != null) {
+              mLayout.addView(infoWindowView);
+              infoWindowView.setOnClickListener(new View.OnClickListener() {
+                @Override
+                public void onClick(@NonNull View v) {
+                  if (onInfoWindowClickListener != null) {
+                    onInfoWindowClickListener.onInfoWindowClick(airMapMarker);
+                  }
                 }
-              }
-            });
+              });
+            }
+          } else {
+            webView.loadUrl(
+                String.format(Locale.US, "javascript:showDefaultInfoWindow(%1$d);",
+                    airMapMarker.getId()));
           }
 
           ignoreNextMapMove = true;
         }
       });
     }
+
+    @JavascriptInterface
+    public void markerDragStart(final long markerId, final double lat, final double lng) {
+      handler.post(new Runnable() {
+        @Override public void run() {
+          if (onMapMarkerDragListener != null) {
+            onMapMarkerDragListener.onMapMarkerDragStart(markerId, new LatLng(lat, lng));
+          }
+        }
+      });
+    }
+
+    @JavascriptInterface
+    public void markerDrag(final long markerId, final double lat, final double lng) {
+      handler.post(new Runnable() {
+        @Override public void run() {
+          if (onMapMarkerDragListener != null) {
+            onMapMarkerDragListener.onMapMarkerDrag(markerId, new LatLng(lat, lng));
+          }
+        }
+      });
+    }
+
+    @JavascriptInterface
+    public void markerDragEnd(final long markerId, final double lat, final double lng) {
+      handler.post(new Runnable() {
+        @Override public void run() {
+          if (onMapMarkerDragListener != null) {
+            onMapMarkerDragListener.onMapMarkerDragEnd(markerId, new LatLng(lat, lng));
+          }
+        }
+      });
+    }
+
+    @JavascriptInterface public void defaultInfoWindowClick(long markerId) {
+      final AirMapMarker<?> airMapMarker = markers.get(markerId);
+      handler.post(new Runnable() {
+        @Override public void run() {
+          if (onInfoWindowClickListener != null) {
+            onInfoWindowClickListener.onInfoWindowClick(airMapMarker);
+          }
+        }
+      });
+    }
+  }
+
+  @Override public void setGeoJsonLayer(AirMapGeoJsonLayer layer) {
+    // clear any existing layer
+    clearGeoJsonLayer();
+    webView.loadUrl(String.format(Locale.US, "javascript:addGeoJsonLayer(%1$s, %2$f, %3$d, %4$d);",
+        layer.geoJson, layer.strokeWidth, layer.strokeColor, layer.fillColor));
+  }
+
+  @Override public void clearGeoJsonLayer() {
+    webView.loadUrl("javascript:removeGeoJsonLayer();");
+  }
+
+  @Override public void getSnapshot(OnSnapshotReadyListener listener) {
+    boolean isDrawingCacheEnabled = webView.isDrawingCacheEnabled();
+    webView.setDrawingCacheEnabled(true);
+    // copy to a new bitmap, otherwise the bitmap will be
+    // destroyed when the drawing cache is destroyed
+    // webView.getDrawingCache can return null if drawing cache is disabled or if the size is 0
+    Bitmap bitmap = webView.getDrawingCache();
+    Bitmap newBitmap = null;
+    if (bitmap != null) {
+      newBitmap = bitmap.copy(Bitmap.Config.RGB_565, false);
+    }
+
+    webView.destroyDrawingCache();
+    webView.setDrawingCacheEnabled(isDrawingCacheEnabled);
+
+    listener.onSnapshotReady(newBitmap);
   }
 }